requirements.txt
google-generativeai>=0.3.0
rich>=13.0.0
pydantic>=2.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
regex>=2024.5.0
tenacity>=8.0.0
PyPDF2>=3.0.0
pytest>=8.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
docling>=1.20.0
python-multipart>=0.0.9